OTA update: ignore OTA partitions above OTA_1
The Fri3D Camp badge only uses and updates OTA_0 and OTA_1, for MicroPython.

Yet higher ones might be present for use by additional apps, such as the retro-go emulator.

Therefore, these should be ignored by the over-the-air update code
when it determines which partition should receive the next update.

This patch was only tested briefly in the MicroPython REPL prompt.
A full integration test with an actual OTA update would be wise.